(PC-17806) refactor(auth): move useLogoutRoutine in helpers folder
