+++ src/features/auth/pages/suspendedAccount/FraudulentAccount/FraudulentAccount.native.test.tsx
@@ -11,7 +11,7 @@ import { FraudulentAccount } from './FraudulentAccount'
 jest.mock('features/navigation/helpers')
 
 const mockSignOut = jest.fn()
-jest.mock('features/auth/logout/useLogoutRoutine', () => ({
+jest.mock('features/auth/helpers/useLogoutRoutine', () => ({
   useLogoutRoutine: jest.fn(() => mockSignOut.mockResolvedValueOnce(jest.fn())),
 }))
