+++ src/features/auth/pages/suspendedAccount/SuspendedAccount/SuspendedAccount.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components/native'
 import { useAccountSuspensionDate } from 'features/auth/api/useAccountSuspensionDate'
 import { useAccountUnsuspend } from 'features/auth/api/useAccountUnsuspend'
 import { useSettingsContext } from 'features/auth/context/SettingsContext'
-import { useLogoutRoutine } from 'features/auth/logout/useLogoutRoutine'
+import { useLogoutRoutine } from 'features/auth/helpers/useLogoutRoutine'
 import { navigateToHomeConfig } from 'features/navigation/helpers'
 import { UseNavigationType } from 'features/navigation/RootNavigator/types'
 import { analytics } from 'libs/firebase/analytics'
