+++ src/features/auth/pages/suspendedAccount/SuspensionScreen/SuspensionScreen.native.test.tsx
@@ -21,7 +21,7 @@ jest.mock('features/navigation/helpers')
 jest.mock('features/auth/context/SettingsContext')
 
 const mockSignOut = jest.fn()
-jest.mock('features/auth/logout/useLogoutRoutine', () => ({
+jest.mock('features/auth/helpers/useLogoutRoutine', () => ({
   useLogoutRoutine: jest.fn(() => mockSignOut.mockResolvedValueOnce(jest.fn())),
 }))
 const mockedUseCurrentRoute = useCurrentRoute as jest.MockedFunction<typeof useCurrentRoute>
